Split manifest handler middleware out of Project
